Split header files
